My first C task

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	char d;
 	float f;
 
-	printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of a int: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("size of a char: %ld byte(s)\n", sizeof(d));
+	printf("size of a int: %ld byte(s)\n", sizeof(a));
+	printf("size of a long int: %ld byte(s)\n", sizeof(b));
+	printf("size of a long long int: %ld byte(s)\n", sizeof(c));
+	printf("size of a float: %ld byte(s)\n", sizeof(f));
 	return (0);
 }
